Fixed turn and added clipLr function
Fixed turn - PIDControl was returning error < threshold instead of abs(error) < threshold, so the turn error of -100 was always returning true, therefore exiting the loop.
clipLR - this is a version of the clip function, but for a drivetrain. This function can decide whether to slow down one side of the drivetrain or speed up the other, making it so that the max speed can be set higher - and the max speed won't be crossed.